+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## 0.0.6 (2014-09-07)
 
-* New validators: `Date`, `Time`, `DateTime`
-* Unit-testing with tox
+* New validators: `Test`, `Date`, `Time`, `DateTime`
 * `Map` supports `in` checks and now works with `In`: [details](README.md#map)
+* Unit-testing with tox
 * Added changelog :)
 
 ## 0.0.5 (2014-09-05)